+++ src/00_vrt.cpp
@@ -1,5 +1,6 @@
 #include <Rcpp.h>
 #include "gdalraster/gdalraster.h"
+#include "gdalapplib/gdalapplib.h"
 
 
 using namespace Rcpp;
@@ -13,4 +14,11 @@ CharacterVector raster_vrt_cpp(CharacterVector dsn,
   return gdalraster::gdal_dsn_vrt(dsn, extent, projection, sds, bands, geolocation, nomd, overview, options);
  }
 
+
+// [[Rcpp::export]]
+CharacterVector raster_buildvrt_cpp(std::vector<std::string> dsn, 
+                               
+                               std::vector<std::string> options) {
+  return gdalapplib::gdalbuildvrt_applib(dsn, options); 
+ }