+++ NEWS.md
@@ -2,6 +2,10 @@
 
 * New function `vector_vrt()` to generate VRT for SQL and/or reprojection. 
 
+* Fix cross-compilation for ARM on universe, thanks to Jeroen Ooms. 
+
+* New function `buildvrt()` as a special-case for -separate from gdalbuildvrt app. 
+
 * `vapour_vrt()` gains 'options' argument, so we can in particular do `options = c("-expand", "rgb", "-ot", "Byte")` to warp
 16-bit integer colour palettes from GTiff to PNG. :)